Tell the player how to quit if they #quit
Despite several minutes of trying, the extended command code was too
excessively daniel_t to get it to work with client-based commands
like saving and quitting. It contains things like flags fields where
some of the bits sometimes are bits and sometimes form an integer,
and where multiple bitfield constants have the same value and the
code only works through coincidence.

I'm not going to be able to get this fixed before 4.3. So let's just
print a message instead, so that players know how to exit.