Correctly handle turning corridor state for interrupted farmoves
To reproduce in the previous revision: farmove, interrupt the farmove
with the client (via keypress, because prefix counts are optimized
for efficiency), move in another direction (your character will move
the wrong way), then enter a zero-time command (causing a desync).

The fix is for do_move to be aware of the difference between an
unfinished occupation last turn, and a new occupation this turn.