Better save file discipline in unwritable states (game over, recovery)
This now guarantees that when a game finishes, all other processes
reading that game will exit or be forced into replay mode, i.e. we
never go through the game over sequence twice. Also, recovering a
game will no longer potentially cause corruption if another process
is using it at the time (they'll do a turn restart when the save
file reaches its previous length or if they try to give a command).

This also adds a #rewind debug command to test all this, and because
it seems generally useful.  The interface for #rewind is currently
not as good as it could be; it shows the desync error message and
leaves the game in replay mode afterwards.  It's better than not
having it, though.