Fix static links on "windowprocs"
So this is what aimake was trying to tell me all this time; I should
have listened to it earlier. The problem is that both the engine and
the client library had a global variable "windowprocs"; if the client
library is a shared library, this is fine (because aimake prevents
shared library namespace pollution), but if it's being linked against
the engine directly, we have two global variables with the same name,
which is disallowed nowadays (due to the use of -fno-common, which
prevents the variables being merged with each other).

This renames the global in the client to client_windowprocs, fixing
everything in a much neater way than the AIMAKE_REVERSE_IMPORT method
we used to use (which is basically a way of telling aimake "shut up,
I know what I'm doing", and I didn't).