Fix a corner case of a corrupted logfile
Fixes #569. Not sure if this actually happened to anyone or if it
was found by reading the code; it only seems to happen as a result
of a file getting truncated within the first three lines and not
at a line boundary, which is not something I've seen in the wild.