Add a command for crashing the client
For debug purposes. At the moment, we have two problems: working
out why it's crashing, and working out why it isn't producing a
crash dump. This allows us to separate the two problems from each
other. Limited to situations where the client has a wizard mode
game open, so as to not impact server play.