Fix off-by-one in buffer allocation in curses_yn_function
The previous code happened to work on Linux due to the order in which
the stack is allocated (not that we should rely on that), but crashes
on OS X, which is better at catching this sort of buffer overflow.

Memcheck can't catch this sort of error. I guess we'll have to wait
for sgcheck to become more mature (it currently can't handle NetHack
4) before we can get automated detection of this happening.