Overhaul handling of water and lava in movement commands
This fixes #536, and at least one vanilla bug (in 3.4.3, break a
fountain while blind, then try to travel into one of the resulting
pools from a distance). This also partially fixes #232, but there
are still outstanding issues there.

Note to future developers: when seenv is used as a boolean, it means
"the player character has seen this square at some point in the
past", not "the player character currently has an accurate idea of
what is on this square". Normally, you should be using mem_bg
instead. I'd go so far as to say that reading seenv except in
connection with wall_angle or code meant to debug wall_angle is
almost certainly a bug (such as the currently unfixed use in
do.c:697 to see whether the character knows what is beneath them).