Fix inventory letter corruption
The trigger is picking up an item when slots 'a'-'z' and 'A'-'Y' are
full, the last item picked up went into slot 'Z' (but that slot is
now empty), and a not-previously-picked-up item is picked up. These
circumstances are specific enough that it's not surprising this bug
wasn't found earlier, but it tends to persist in a save file once it
has been.

To fix an affected save on this new revision, drop your entire
inventory using DAa on a blank map square, then pick it up again.
There's a save compatibility shim that fixes the internal inventory
letter state, but the drop-and-pick-up is needed to fix it on the
individual items.

Fixes #573.