Rewrite the wield code
Two big (and intertwined) changes here. One is that wielding a
two-handed weapon can now displace shields (when cblock is set),
and likewise wearing a shield can displace a two-handed weapon. The
other is that auto-wielding tools that must be held to apply them
now follows almost the same codepath as a regular wield, rather than
being special-cased. The combined effect is that it's possible for
multiple turns to be spent before you even get a direction prompt
when, e.g., digging, so the code needed some restructuring to account
for that.

This also fixes a rather awkward bug in which equipping a two-handed
weapon with 'w' would succeed despite a worn shield; there was a
reversed test.

Fixes #530.