Move windows onscreen if they wouldn't fit
The code was previously trying to draw windows larger than the
inventory window entirely within the inventory window, which of
course doesn't work. So the window extended offscreen to some
extent, meaning that libuncursed refused to draw it to avoid memory
corruption.

This fixes the bug in nhcurses (#491), and fixes libuncursed to
report ERR rather than OK in this situation (not that nhcurses was
paying attention to the return value anyway).