Fix many of the lingering pluralisation issues
The underlying problem is that the vanilla (non-grammartree) object
naming code produces a singular name (e.g. "pair of leather gloves")
for grammatically plural objects (there are two gloves in each pair);
the code assumes this in many cases (e.g. for buying an object, the
messages are "6 zorkmids for this small mimic corpse" or "6 zorkmids
per small mimic corpse" in singular and plural; there is no format
pattern for "10 zorkmids for these leather gloves", so we need to
generate "10 zorkmids for this pair of leather gloves").

An earlier commit (the Great BUFSZ Purge) accidentally removed the
"pair of", but only when the gloves were formally identified. This
made it absolutely impossible to pluralize correctly in all cases,
and attempts to hack around this via changing is_plural() naturally
didn't work.

That said, being able to generate just "leather gloves" by itself
would be nice, but that's likely going to have to wait for
grammartree.