Fix for ticket #620, Monsters sometimes don't use traps defensively when they should.
This change allows monsters to use a teleport trap as an escape route if it is in a direct line with the player but not actually on the player's tile.

The bug was discovered by examining a non-public commit log, but the fix is obvious and straightforward once the matter is considered.