Upgrade aimake (reduced memory use)
This reduces memory usage for a clean build (which seems to be the
memory-heaviest mode of operation) from 734 metric megabytes to
437 metric megabytes. Not sure if this is enough of a saving for
the people who were having trouble building to build it now. If it
is, this fixes (or at least downgrades) #442.

437 MB is clearly much too much for a build system to be using (I
remember the days when 640KB had to suffice for everyone...), so I
may look for further memory savings in the future.