Fix crash in makeplural()
I'd left it a little too close to the vanilla code, which was for
some reason trying very hard to not do subtractions. This meant
that there were some pointer-into-deallocated-memory bugs
(technically, mortalized rather than deallocated, but the memory
nonetheless had the wrong contents, which is why it was
segfaulting).