A start on a new tileset compiler
This isn't hooked up to the build system yet (as in, aimake builds
it anyway, but doesn't know how to build the tiles). Currently it
is unable to produce binary-format tiles or images; thus, it can
currently only do the job that import_tileset previously did, and
some of the job of convert_tileset. Eventually, I hope to implement
the rest of the functionality of convert_tileset, and export_tileset
too.