Implement embedding tilesets into an image
.nh4ct is now a viable alternative to .txt for long-term storage of
tile images (although .txt is probably better for the purpose); it's
now also usable as a format for the game core to start reading.

Before we can have the new tileset system, three things are needed:
reading the .txt image format (currently unimplemented), hooking
all this up to the build system, and getting the client to understand
the new tileset format.