Terminal handling fixes
Fixes #506.

There are two main changes here. One is to fix a bug in ioctl
discipline when stdin and stdout refer to the same underlying file
(this is particularly big, because stdin and stdout /usually/ refer
to the same underlying file!). The other is a brute-force string
termination (8KiB of NUL characters) every time we clear the screen,
to help out terminals like dvtm which get stuck in a non-default
parse mode.