Refactor deferred gotos into turnstate.
This is part of me trying to cleanup for levels branch, but this fix should go
on master. Note that 24 bytes of the you structure become unused, but they are
left in the save file so as to avoid breaking save compatibility.

The SAVEBREAK tag in the comment is new; the intent is that when do we create a
save-breaking branch, we will remove this.