Don't override underlining except on monsters and cmaps
This allows future implementation of, e.g. underlining item piles
that contain more than one item. At least, freeing up the underline
bit for future expansion is better than not doing it.

A side effect of this is that when an item is dropped on a long
staircase, its underline shows through. This might or might not be
considered desirable behaviour.