Fix a lot more warnings.
This also uses a bleeding edge (read: I haven't submitted the patch yet) clang
attribute to silence some warnings on enumerators. Because of clang's feature
test macros, it will compile correctly on clangs without it, so there's no harm
in including it other than the possibility that the patch may change before it
lands.