Copyright status updates for tilesets
Wow, the situation is complex. This is what starts happening when you
start working with image licenses as well as code licenses.

This corrects the copyright status on DawnLike (CC-BY-SA, except for
tiles from DawnHack which are CC-BY), and on RLTiles (which has a
custom attribution license; it claims to be public domain, but this
is dubious given the requirement for attribution). It also updates
all the copyright/history documentation to describe the copyright
status and contributor list. (I still need to ask Raz what name
he/she wants to be credited under; using "Raz" for now.)