Upgrade aimake (rpath-related fixes)
This gets rid of one Lintian warning. We're now down to five at
"warning" level or higher. Two are correct warnings about the fact
that we're using one package when we should be using at least three
(nethack4, libuncursed, libuncursed-dev) and arguably also
nethack4-data, nethack4-tiles, and libuncursed-sdl (but six
packages would be a huge pain for users unless they were all in the
package manager). One is the fact that this isn't aimed at any
version of Debian in particular; Lintian correctly observes that
this would prevent it from being uploaded to Debian, but we're just
planning to serve these via the website. One is the fact that the
license is duplicated between NetHack's data files and the copyright
files; this is probably unavoidable because we need to be able to
show it in-game, and it isn't in common-licenses. And one is the
fact that we're installing a directory at 0775 root/games, which is
not a problem, and likely a Lintian bug that it's being reported as
one.