Get known folders working correctly on Windows
Now we can actually do something other than dump everything in the
same directory, yay!

This will break the Linux server install, as it hasn't yet been
updated to the new API. The aimake documentation hasn't been fully
updated either. So this is something of a partial commit; making it
now over on the Windows end as the easiest way to transfer the data
to Linux.