Convert copyright.binary to RTF
This is intended for Windows, because Linux users typically care more
about the source than the binary. RTF is what installers want, and it
lets me make it look somewhat prettier (if less machine-readable). I
also put a summary at the start, for the benefit of people who
actually read this sort of thing (like me).