Ticket #601, xlogfile should have a field for birth options, first draft.
This adds a function to encode birth options.  Currently it handles
elbereth, rogue, seduce, bones, permablind, and permahallu; I am
thinking about adding polyinit (just a yes/no bit, not the details)
but will do that in a separate commit.  The function is used by
write_xlentry(), placing the new field right after the conduct field.
It is formatted in a similar fashion to the conduct field.