Fix a desync with shopkeeper corpses.
Basically the problem was in the assumption that one could safely memcpy a
monster around. This worked fine except that eshk had this amazingly dumb design
of a pointer into its own data structure. I've replaced this with a boolean flag
which does the job in a much cleaner fashion.

No, I don't know what the actual semantics of bill_inactive are, and I don't
care, quite frankly.

The reason it's 4 bytes and backwards is because, coincidentally, it would
always be saved as 0 whenever bill_p was equal to &bill[0]. that means that we
can simply translate it to an inverse boolean flag and no extra shims are
required.

I'm evil.