Replace the random number generator
This change breaks midturn save compatibility; if your game was
halfway through a turn before this commit, there's a chance that you
might be asked to rewind to the start of the turn. I couldn't keep
the old engine around due to the copyright infringement.

Sadly, this doesn't seem to have helped save file size much; we're
down to 24 bytes for an action (walking into a wall) that changes
the RNG and has no other effects, but most diffs are way longer
than that, so the RNG probably wasnt at fault in the first place.
It's an improvement, but not one that really matters.

However, there are various other possibilities that this opens up.
For instance, it's now possible to attach specific RNGs to specific
actions (there are 209 of them), such as whether djinn give a wish.
I plan to add specific RNG attachments to random effects that have
a large impact on game balance, and also some method of specifying
the initial RNG seed. (The seed of a game is written to the xlog
once it's over, so you could "reproduce" a game.)

This also fixes one of the known causes of gamestate desync after
game over (program_state.gameover now permits unrestricted RNG use).

Currently broken: RNG seeding on Windows.