Write failures to open files to the panic log
This will hopefully help me debug what's up with the dumplogs in
server play (where there are remarkably few safe places to write
debug information). As a bonus, the panic log happens to be an
appropriate place for this sort of information anyway, so I'm
commiting this change as a permanent improvement rather than just
using it as a temporary debugging aid.