Ticket #728: make tinnable() do a sanity-check whether corpse is indeed a corpse.
It seems clear from the argument name that the original intention was
that only corpses should ever be passed as an argument, but in
practice the function was being called to determine what _objects_
were reasonable to apply the tinning kit to, which resulted in being
able to tin things that weren't even food.

It has been suggested that it should also be possible to tin other
things, e.g., kelp.  (Fruits and vegetables seem likewise obvious.)
I have not attempted to implement that at this time, since it would
be much more involved and would add kinds of tins that do not exist
in vanilla.  Perhaps we could revisit that later, but for now this
just (hopefully) fixes the bug.

Lightly tested:  with this change, I can still tin a corpse off
the floor, but with other objects (e.g., rations, scrolls) on the
floor at my feet, the game tells me I have nothing to tin.

I have not done a full audit of every place tinnable() is called.