AI improvements for swarms
This contains changes to help swarms keep their cohesion (not 100%
reliable, but it's better than previous versions). It also contains
optimizations for very large groups of monsters (e.g. pudding farms):
we run only a minimalist AI for monsters in the middle of the swarm,
to cut down on CPU and save file size. Some experimentation shows
that this is almost impossible to notice in-game, and the main
visible effect is that pudding farming becomes a little less reliable
because the patterns in which the puddings turn up have become less
predictable. (Not that this is likely to make pudding farming
impossible or anything, although some pastescripts may starve to
death without modification.)