A new save diffing algorithm that produces better output
I'm pretty happy with the diff format, at this point (although I
left a two-byte header so that I can change it if necessary). The
encoder, on the other hand, is not perfect, and produces peephole-
optimizable output in several circumstances. (This is because it's
currently using a one-pass algorithm.)

Still, this is better than before; on several occasions, this is
being output uncompressed, because zlib couldn't compress it any
further. (zlib is still useful when compressible data is added to
the save; the added data appears verbatim in the diff, giving zlib
a chance to compress it.)