Let pets use the improved monster pathfind code
It's nice to remove some nonworking code and replace it with calls to
code that already exists; you get a functionality improvement, cleaner
code, and it's less work to maintain it.

The main user-visible change here is that pets are less likely to get
stuck in corridors trying to follow the player over long distances
(e.g. one side of the level to the other). There are basically two
reasons why a pet might not follow in 3.4.3: because it doesn't want
to, or because it can't figure out how. This patch eliminates the
second problem, while leaving the first unchanged.