Ticket #240: differentiate "Unarmed" status.
The status line now shows "Unarmed" (in purple) only if the player is
wielding nothing or a non-weapon or missile (e.g., dart) or a polearm
while not riding; "Dig" (in brown) if wielding a pick-axe or mattock;
"Ranged" (in green) if wielding a launcher; and for completeness
"Melee" (in gray) if wielding a melee weapon.  Tested at 80x24, at
100x35, and at 152x52, all in combination with also being overloaded,
hallucinating, stunned, blind, confused, and satiated, all at once.
At 100x35 this can cause the turn count not to be shown, particularly
if the player is carrying several digits' worth of gold.  That seems
like an acceptable outcome for such a pathological case.