When loading a bones level, update the dlevel.
When loading a special level, it's possible to load it in a different place from
where it was originally placed in the original dungeon. As a consequence, the
level will think it's in the wrong place in the levels array and all sorts of
Bad Stuff is liable to happen, and surprsingly not too much did.

In order to maintain compatibility, this will only apply to bones levels loaded
with the new code. The noticeable gameplay effect (besides fewer bugs) is that
when a special level is moved, it will now generate monsters at its actual depth
in the dungeon, not the depth at which it was originally. This makes
bones-Minetown a little less frightening, I guess, since it's not going to
accidentally be several levels deeper than it appears to be.

The upshot of this is that #784 is fixed, because the shopkeeper was comparing
if the level it's on was the same as the shoplevel, and this was *not* the case
due to the incorrect dlevel in the level itself. So the shopkeeper knew where he
was; just that the level itself didn't and as a result, the shopkeeper thought
he wasn't home.