Monster AI audit: mux/muy, invisibility, and displacement
Fixes #745, and probably also #759.

This changes the internal format for muxy (i.e. mux/muy) to the same
COLNO/ROWNO sentinels that are now standard elsewhere, and audits
every use of them throughout the code.

This also contains some AI bugfixes (e.g. in 3.4.3, a monster can
attack thin air when attacking an invisible player, and actually be
attacking an occupied square). The largest change is a new
abstraction that allows monsters to attack /map squares/, rather than
players or monsters; this lets us handle all (well, most, some are
unimplemented) effects of attacking a displaced image, and simplifies
the code somewhat. An example of a monster behaviour that wouldn't
happen before this commit: monsters with breath weapons can aim them
along lines that don't hit the player, if confused by displacement.
This change actually simplifies the code, which is nice, as well as
making it more player/monster symmetric, which is very nice.

The behaviour for invisibility isn't as good as I'd like. What's
happening right now is that as soon as you turn invisible, monsters
don't even dare to guess at your position. So they follow your
tracks, and typically bump right into you while tracking. This is
realistic but a bit repetitive; in a future commit, I'll aim to
implement a "gray I" effect on monsters, so that they'll attack
the player's remembered position if the player moves away.