Rework the keymap submenu patch
This uses a more consistent interface for mass rebinding: you choose
what keys you want to rebind first, and then what actions you want
to give them. Direction commands are also included as an appendix to
the list of game commands (because it's weird for "run in a given
direction" and "run as a prefix key" to be on diferent menus).