Warnings fixes
I just upgraded gcc. One of these is really bad, and could prevent
save files porting between systems with different struct padding
(we were reading past the end of an array in the middle of struct
you, then writing the data back past the end of the array; this is
undefined behaviour but will likely work in practice so long as
the padding stays the same).