Fixes in pet item drop handling
Fixes #718. The problem was that a pet wouldn't drop a weapon that
was better than its currently wielded weapon, if its currently
wielded weapon isn't welded. The intent was that the pet would swap
weapons and then drop the bad one. The problem is, there was no
check to ensure that the pet could wield weapons in the first place,
so domestic pets would just keep the weapons forever.

The reason for the large diff is that I've decided that when adding
new simple accessor functions on monsters, I'm going to const-correct
the monster pointer. This way, the code is going to become slowly
const-correct for monsters, a bit at a time. The accessor I added
here was perhaps a little less simple than I expected.