Implement an Alt-is-ESC mode for laggy network connections
I was considering doing this in libuncursed, but the semantics are
much clearer doing it in the client instead: "whenever someone
presses Alt-letter, interpret it as ESC letter". The idea is that
you can turn this on on a laggy network connection to prevent ESC
letter being misinterpreted as Alt-letter, at the cost of the Alt
key no longer working properly. (For bonus points, you can then
configure your terminal to Alt-is-Meta mode, because Meta still
works.)