Fixes for clear_path and engulfers
This is basically an audit of assumptions that "player's square =
player": sometimes it isn't (it can be an engulfing monster), and
sometimes it is but the couldsee() data is unavailable. Fixes #824,
as well as a bug where lights in the players' inventory wouldn't
emit light after doing a full screen redraw.