Greatly simplify color-mapping code
This removes the support for 256-color and 16-color curses color
modes, because those don't work with libuncursed (because it's
impossible to remotely detect that they'll work). colormap.conf is
likewise removed on the same basis. The intended replacement is the
palette-mapping code in the options menu, which works along an
entirely different axis (and has the bonus of being settable
in-game, although it could do with a greater selection of palettes).