Add copyright information on every file in the repository
Apart from NetHack4.nsi, which was unused. This involved inventing
a comment syntax for rumors.fal/rumors.tru, and some mindbending
recursive legalese for the copyright statuses of COPYING and
binary-copyright.pod.

I've just noticed that the license for the license is defined as
being under the same license as itself.  I'm going to leave it like
thatfor now, though; it explains why it's legal to copy it (because
the requirement to preserve the license implies that you're allowed
to preserve the license), and tautologies are fun.