Initialize unused member of struct dflags, to avoid compiler warning.
Because the member is unused (at least, I assume it is unused, since
it is literally named "unused" and was not being initialized with the
other members of the struct and had been that way since NitroHack),
the value in principle does not matter at all, and so its being
uninitialized was presumably not causing any actual problems.  So
this fix is, strictly speaking, unnecessary.

But I'm just neurotic enough that compiler warnings occasionally bug
me enough that I do something about them, especially if they're ones
that are easy to do something about.  I arbitrary chose to initialize
the member to 0.  (The only other possible value would be 1, since
it's a single unused bit.)