Some compatibility fixes for -beta1 saves
Probably mostly fixes #830. The -beta1 save I have for testing #830
on is insane on the replay, though; it shows someone levelporting out
of bounds. Their actions are consistent with those of a player who
accidentally levelported out of bounds (try to dig out, then try to
teleport out), so I'm going to assume that that is what actually
happened and the replay is replaying correctly. -beta1 bugs are
weird.