Color fixes for the HP bar
In particular, when the foreground is yellow, the background is
brown, and we draw yellow foreground on brown background, rather than
black foreground on brown background, for digits that are on top of
the bar. (This is consistent with color handling on the map.) I think
this looks slightly better than the old method.