Remove code duplication for changes in monster attitude
When changing monster attitude, we nearly always want to redraw, and
we frequently want to change the monster's alignment yield. Let's
bundle this all up into one function rather than having tens of copies
all over the code.

Also, an unusual sight: a code duplication fix that actually makes the
code shorter (even if I had to do some whitespace fixes in the process
to get away with that). Normally, adding comments and the like means
that the code gets longer in total.

Fixes #150.