Allow some game directories to be omitted
This is used by the testbench to not produce dump files or bones
files. It might be useful elsewhere, too.

The special token $OMIT is used for this, mostly to prevent the
mechanism being triggered by accident (because it disables warnings
about failures to find files on the basis that that's intended
behaviour).