Handle strength drain while polymorphed correctly
The previous implementation was inconsistent. Some sources would
kill a polymorphed character if their hitpoints dropped below 0;
others didn't handle the codepath (leading to a rehumanization at
end of turn in 3.4.3 because it's HP death, and an impossible() in
NH4). This change consistently makes the situation a rehumanization,
with an appropriate death message, immediately upon taking strength
drain that drops your HP below 0 while polymorphed.

Found by the testbench.