More sensible behaviour for a spellbook being cursed while you're reading it
The previous codepath called into part of the cursed spellbook
effects code, but not the whole thing. This meant that you'd have
no helplessness, you'd memorize the spell, you'd get the bad result
that comes before the helplessness, and the book would be destroyed
with the explosion result, and incorrectly marked in-use with the
contact poison result (causing a message and destruction of the
spellbook the next time you loaded a save in 3.4.3, or a desync in
4.3). The new behaviour is that the instant the book becomes cursed,
you're left helpless for a few turns, and you don't learn the spell.
(The helplessness duration is 4 + d5 turns; reasonably low, because
you may well be in combat; slightly higher than might otherwise be
expected, because if you are in combat you should have stopped
reading, and for Rodney-curses, this needs to be large enough to be
relevant.)

I suspect the contact-poison desync wasn't noticed before because
in 3.4.3, the cause and effect are mostly unrelated, and it requires
a sequence of coincidences.

Found by the testbench.