Fix bugs in demon generation
Found by the testsuite.

This bug possibly also exists in 3.4.3, but would be almost impossible
to reproduce. You'd need at least to generate a morgue after the
extinction of all normal demons, and there are also constraints on
depth (dlvl 12 seems to be a good place; the testsuite found this
problem there twice). Even then, it might not happen; you'd need to be
(un)lucky.