Don't crash when a piranha steps on a teleport trap with no water on the level
Found by the testsuite.

This pretty much comes down to the fact that it's playing so many
games; it wasn't specifically testing for this. It just spent a long
time on a level with a polytrap and a teletrap right next to each
other, and that happened.

Also, I felt sorry for the piranha, until it killed my character
in the game where I was testing this.

Unlike the other bugs the testsuite has found, this one is actually
possible to deterministically reproduce, if you know what you're
looking for.

This isn't a problem in 3.4.3 because it has a 1/13 chance of
placing an aquatic monster on land anyway. That change was removed
for 4.3-beta2 so that monster placements on a newly created level
would be determinstic.