Change the 'graphics' option (enum) to 'tileset' (string)
Also contains a number of bugfixes to do with string-valued options,
in order to get this working or that I noticed while debugging it.

The idea is that this makes it possible for users to use their own
custom tilesets.  The UI isn't great, I'll fix it in a future commit.