Preliminary livelog implementation.
This implements a livelog for NH4, i.e., a log of important game
events as they occur.  Logged events include receiving information
from the Oracle for the first time; killing (or otherwise ensuring the
death of) Medusa, your quest nemesis, Vlad, Rodney, the High Priest of
Moloch, Demogorgon, or Croesus for the first time; killing the ghost
of a former player from a bones level; making a wish; being crowned;
performing the invocation ritual; entering the first or last Elemental
Plane; or having your life saved by an amulet of life saving.  The
most spammable of these is wishing, but players tend to like seeing
wishes announced in IRC channels, and in any event wishing is nowhere
near a spammable as death (which shows up in the xlog).

The bones ghost kill logging feature doesn't quite work yet, according
to my testing.  Also, it breaks save compatibility, so it's wrapped in
save file compatibility break.