Optimizations in libuncursed
This makes the code a little faster in typical use (where the total
area covered by refreshed windows is greater than the area of the
screen; this is true for NetHack 4 and probably most programs that
do enough refreshing for the difference to be noticeable). It also
makes the code a lot simpler.