Make failing to displace a pet cost a turn
Not costing a turn is obnoxious interface-wise (you can't get past
a pet in a shop just by repeatedly moving into it, because it has
no opportunity to get out of the way), and because pet displacement
is based on a random roll, means that outside shops it's possible
to spam moves into the pet until they work, with no penalty for
trying. This mechanic thus works much better as time-consuming.