Livelog improvements:  add some more useful fields.
First I wanted the birthdate and starttime fields so I can put
xlogfile and livelog entries together and sort them cronologically.
Then I realized I wanted mode so I can filter on it.  I noticed that
character is the same information as xlogfile's charname but with a
different name, so I renamed it to match.  Then I ended up adding
version and xplevel, as well as role, race, align, and gender.