Don't crash when teleporting monsters on full levels
This fix is for all the cases where leaving a monster in place would be
unproblematic (mostly cases which either have a random factor anyway, or
where the monster stays put on no-teleport levels). There are still some
cases which need more thought, e.g. preventing a bones ghost and steed
occupying the same square. (In many of these cases we'll have to kill a
monster to make room.) I left those cases as impossible()s for now.