Panic on trying to run AI for off-level monsters
This is happening in Morwen's pudding farming game, after the
bugfixes. Add a panic to make the problem a little easier to identify;
this doesn't fix the bug, but makes it clearer that it's happened (most
of the monster AI runs just fine with monsters like this which are
slightly offlevel, because it's handled the same way as the monster
being at an edge).