Monster probabilities for G_INDEPTH disabled that match 3.4.3 better
I was using a fixed rejection chance of 1/2, whereas 3.4.3 applies
the rejection chance once for every monster strength band that would
be out of depth at half the dungeon level. This also fixes the
implementation of frequency stealing (replacing it with a rejection
chance rather than a frequency adjustment; the old code was highly
different from 3.4.3 at pretty much every XL, the new code matches
it at most XLs and is strictly more lenient at very low XLs).

Fixes #878, probably. (This sort of bug is almost impossible to
test for without lots of anecdotal data, so it's hard to tell whether
it's fixed or not.)