Steer each spell towards a specific letter by default
It'd be nice if we could do this for everything, but we can't;
there are way more than 52 possible inventory items. However, there
are less than 52 magical abilities, so we can steer the spells
towards letters (not just the supernatural abilities, like the
previous commit does). We have the ability to assign spell letters
out of order now; may as well use it.