Handle the cases of migration to a full level that 3.4.3 handles
Not fixed, but now fixed for MIGR_RANDOM at least, which caused
problems for at least two players. (We kill the arriving monster
and drop its items, as in 3.4.3.) This was a mistake from the
COLNO/ROWNO conversion.

This also adds a panic in the failing case, which is better than
leaving the monster at invalid coordinates.