Ticket #876, attempt to fix the #monster crash
When polyselfed into a spell-casting monster, such as a titan or
arch-lich, if the player attempted to #monster, there was a segfault
when mmspell_would_be_useless attempted to call m_mx() and m_my()
on a NULL pointer, which had been passed in by castum().

This is my first attempt at fixing it and works by having the
mmspell_would_be_useless function expressly check whether mdef is true
each time it uses it.  This does avoid the segfault.  (Instead, the
player now gets a message, "You have no spells to cast right now.")
Whether it is the /correct/ fix, I'm not sure.

Cherry-picked from 88907b1a5bbcc473198b2792894a87046ba79f57