Rewrite uim-related code
Instead of having bits of figuring out the user intent scattered
all over the place, they're now centralised: we have one function
for figuring out user intent based on their knowledge and the uim,
and then try to apply that intent in whichever direction they
chose (possibly randomized by stun/confusion).

Note that this has a gameplay-affecting sequencing change. If
you're stunned and try to move into an empty space, then even if
the direction of your action is randomized, it'll still be a move
(so, say, you might bump into a monster). Likewise, if you try to
attack, you'll end up attacking in a random direction (and might
end up hitting a boulder or some stairs instead, for example).

This also means that the "attackonly" uim can be removed; it was
previously only needed to make the internals work. It's been
replaced by an "attackhostile" uim that attacks only hostile
monsters, and doesn't even prompt for peacefuls; there seems to
be some demand for that.

Things that might need changing: with these changes, the only way
to hit a tame monster at range with a polearm is to go into the
options and set movecommand up to indiscriminate or forcefight
(because other movecommand values require a prefix 'F' to hit tame
monsters, and that prefix only works on direction commands, not on

Fixes #894, #911, #920, #925, and some of #915 and #862.